Move axis_world_coords test to own module.
